app/pyproject.toml
[tool.poetry]
name = "japan-ifc-building-confirmation"
version = "0.1.0"
description = ""
authors = ["ktaroabobon <ktaroabobon@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
python-ifcopenshell = "^1.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
